// frontend/frontend/src/types/Bowler.ts
export interface Bowler {
  bowlerFirstName: string | null;
  bowlerMiddleInit: string | null;
  bowlerLastName: string | null;
  teamName: string;
  bowlerAddress: string | null;
  bowlerCity: string | null;
  bowlerState: string | null;
  bowlerZip: string | null;
  bowlerPhoneNumber: string | null;
}